Update web-hooks-push.md
The Wikipedia link we include in this RCP, as well as many other sources, states "webhooks" as one word instead of "web hooks" with a space, so I have made that change here along with other tiny formatting updates.

Why I think this matters: Unified phraseology is good for marketing, which is my end of the pool at RESO. For example, I used to see WebAPI, WEBAPI, WEB API, Web API and other iterations. With effort, I now mainly see only Web API, and I believe it helps the term resonate with a wider audience. I want the same for webhooks and all RESO standards.

I stopped short of changing the link from web-hooks-push.md to webhooks-push.md, for fear that there are links that would break in the wild, but my lizard brain wants to do that, too, haha.

I'm not great at GitHub, but I think I know enough to make this a branch or proposed change rather than committing directly to the main branch. Walk before you run!

It is my understanding that if the changes are acceptable that they will be merged into the main branch by someone closer to the original proposal. But if the changes look fine, I believe I have rights to commit directly to the main branch. But I wanted to exhibit "polite behavior" before doing anything like that. (See what I did there?)